+++ CHANGELOG.md
@@ -44,6 +44,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1300](https://github.com/alleslabs/celatone-frontend/pull/1300) Apply Initia's constants from initia react widget to username and provider
+- [#1302](https://github.com/alleslabs/celatone-frontend/pull/1302) Convert more text to sentence case
+- [#1298](https://github.com/alleslabs/celatone-frontend/pull/1298) Fix tooltip click capture on TokenCard
 - [#1296](https://github.com/alleslabs/celatone-frontend/pull/1296) Convert title to sentence case
 - [#1294](https://github.com/alleslabs/celatone-frontend/pull/1294) Upgrade React & ESLint version
 - [#1293](https://github.com/alleslabs/celatone-frontend/pull/1293) Improve useOpenTxTab and useOpenBlockTab