+++ CHANGELOG.md
@@ -43,8 +43,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bug fixes
 
-- [#1306](https://github.com/alleslabs/celatone-frontend/pull/1306) Fix validators list separate divider UI
-
 ## v1.11.0
 
 ### Features
@@ -67,6 +65,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bug fixes
 
+- [#1306](https://github.com/alleslabs/celatone-frontend/pull/1306) Fix validators list separate divider UI
 - [#1305](https://github.com/alleslabs/celatone-frontend/pull/1305) Remove Initia Terms, update initia widget to v1.3.5
 - [#1297](https://github.com/alleslabs/celatone-frontend/pull/1297) Fix select input floating label